requirements.txt
appdirs==1.4.3
attrs==19.3.0
Automat==20.2.0
beautifulsoup4==4.9.0
boto==2.49.0
boto3==1.12.39
botocore==1.15.39
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9
cssselect==1.1.0
dateparser==0.7.4
docutils==0.15.2
feedfinder2==0.0.4
feedparser==5.2.1
future==0.18.2
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.6.0
incremental==17.5.0
jieba3k==0.35.1
jmespath==0.9.5
joblib==0.14.1
lexrank==0.1.0
logger==1.4
lxml==4.5.0
newspaper3k==0.2.8
nltk==3.4.5
numpy==1.18.2
opencv-python==4.2.0.34
pandas==1.0.3
parsel==1.5.2
path==13.1.0
path.py==12.4.0
patsy==0.5.1
Pillow==7.1.1
Protego==0.1.16
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
PyYAML==5.3.1
queuelib==1.5.0
regex==2020.4.4
requests==2.23.0
requests-file==1.4.3
s3transfer==0.3.3
scikit-learn==0.22.2.post1
scipy==1.4.1
Scrapy==2.0.1
sentencepiece==0.1.85
service-identity==18.1.0
six==1.14.0
smart-open==1.11.1
soupsieve==2.0
SQLAlchemy==1.3.15
statsmodels==0.11.1
testfixtures==6.14.1
tinysegmenter==0.3
tldextract==2.2.2
tqdm==4.45.0
Twisted==20.3.0
tzlocal==2.0.0
uritools==3.0.0
urlextract==0.14.0
urllib3==1.25.8
w3lib==1.21.0
zipp==3.1.0
zope.interface==5.1.0
